chore: rename variables
