+++ .storybook/manager.js
@@ -1,7 +1,7 @@
 window.STORYBOOK_GA_ID='UA-159327743-4' // Google Analytics
 
 import { addons } from '@storybook/addons';
-import SpaceOneTheme from './SpaceOneTheme';
+import SpaceOneTheme from './CloudforetTheme';
 
 addons.setConfig({
     theme: SpaceOneTheme,
